+++ app/cells/folio/atom/text_cell.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
-class Folio::Atom::TextCell < FolioCell
+class Folio::Atom::TextCell < Folio::ApplicationCell
 end