+++ app/cells/folio/console/form/header_cell.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class Folio::Console::Form::HeaderCell < FolioCell
+class Folio::Console::Form::HeaderCell < Folio::ConsoleCell
   def translations
     cell('folio/console/nodes/translations', model.original, as_pills: true)
   end