+++ app/cells/folio/console/form_footer_cell.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class Folio::Console::FormFooterCell < FolioCell
+class Folio::Console::FormFooterCell < Folio::ConsoleCell
   def class_name
     'folio-console-form-footer--static' if options[:static]
   end