+++ app/cells/folio/console/menu_tree_cell.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
-class Folio::Console::MenuTreeCell < FolioCell
+class Folio::Console::MenuTreeCell < Folio::ConsoleCell
 end